+++ pkg/api/password/doc.go
@@ -1,0 +1,2 @@
+// Package password contains the service that handles passwords
+package password
