+++ pkg/api/store/user_test.go
@@ -1,4 +1,4 @@
-package mysqldb_test
+package store_test
 
 import (
 	"fmt"
@@ -6,11 +6,11 @@ import (
 	"time"
 
 	"github.com/jinzhu/gorm"
+	"github.com/stretchr/testify/assert"
 
-	"github.com/johncoleman83/cerebrum/pkg/api/user/platform/mysqldb"
+	"github.com/johncoleman83/cerebrum/pkg/api/store"
 	"github.com/johncoleman83/cerebrum/pkg/utl/mock/mockdb"
 	cerebrum "github.com/johncoleman83/cerebrum/pkg/utl/model"
-	"github.com/stretchr/testify/assert"
 )
 
 var (
@@ -138,7 +138,7 @@ func TestCreate(t *testing.T) {
 		t.Error(err)
 	}
 
-	udb := mysqldb.NewUser()
+	udb := store.NewUser()
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
@@ -171,13 +171,14 @@ func TestView(t *testing.T) {
 		expectedData *cerebrum.User
 	}{
 		{
-			name:        "User does not exist",
+			name:        "User should not not exist and not return error",
 			expectedErr: true,
 			id:          1000,
 		},
 		{
-			name: "Success",
-			id:   3,
+			name:        "Success",
+			id:          2,
+			expectedErr: false,
 			expectedData: &cerebrum.User{
 				Email:      "tomjones@mail.com",
 				FirstName:  "Tom",
@@ -187,10 +188,9 @@ func TestView(t *testing.T) {
 				CompanyID:  1,
 				LocationID: 1,
 				Password:   "newPass",
+				Token:      "asdf",
 				Base: cerebrum.Base{
-					Model: gorm.Model{
-						ID: 3,
-					},
+					Model: gorm.Model{ID: 2},
 				},
 			},
 		},
@@ -203,11 +203,13 @@ func TestView(t *testing.T) {
 		t.Error(err)
 	}
 
-	udb := mysqldb.NewUser()
+	udb := store.NewUser()
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
 			user, err := udb.View(db, tt.id)
+			fmt.Println(err)
+			fmt.Println(tt.expectedErr)
 			assert.Equal(t, tt.expectedErr, err != nil)
 			if tt.expectedData != nil {
 				if user == nil {
@@ -229,44 +231,32 @@ func TestView(t *testing.T) {
 	}
 }
 
-func TestUpdate(t *testing.T) {
+func TestFindByUsername(t *testing.T) {
 	cases := []struct {
 		name         string
 		expectedErr  bool
-		usr          *cerebrum.User
+		username     string
 		expectedData *cerebrum.User
 	}{
 		{
-			name: "Success",
-			usr: &cerebrum.User{
-				Base: cerebrum.Base{
-					Model: gorm.Model{
-						ID: 3,
-					},
-				},
-				FirstName: "Z",
-				LastName:  "Freak",
-				Address:   "Address",
-				Phone:     "123456",
-				Mobile:    "345678",
-				Username:  "newUsername",
-			},
+			name:        "User does not exist",
+			expectedErr: true,
+			username:    "notExists",
+		},
+		{
+			name:     "Success",
+			username: "tomjones",
 			expectedData: &cerebrum.User{
 				Email:      "tomjones@mail.com",
-				FirstName:  "Z",
-				LastName:   "Freak",
+				FirstName:  "Tom",
+				LastName:   "Jones",
 				Username:   "tomjones",
 				RoleID:     cerebrum.AccessRole(100),
 				CompanyID:  1,
 				LocationID: 1,
 				Password:   "newPass",
-				Address:    "Address",
-				Phone:      "123456",
-				Mobile:     "345678",
 				Base: cerebrum.Base{
-					Model: gorm.Model{
-						ID: 3,
-					},
+					Model: gorm.Model{ID: 2},
 				},
 			},
 		},
@@ -275,32 +265,88 @@ func TestUpdate(t *testing.T) {
 	container := mockdb.NewMySQLDockerTestContainer(t)
 	db, pool, resource := container.DB, container.Pool, container.Resource
 
-	if err := mockdb.InsertMultiple(db, superAdmin, cases[0].usr); err != nil {
+	if err := mockdb.InsertMultiple(db, superAdmin, cases[1].expectedData); err != nil {
 		t.Error(err)
 	}
 
-	udb := mysqldb.NewUser()
+	udb := store.NewUser()
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
-			user := &cerebrum.User{}
-			if err := db.First(user, tt.usr.ID).Error; err != nil {
-				t.Error(err)
-			}
-			tt.expectedData.CreatedAt = user.CreatedAt
-			tt.expectedData.LastLogin = user.LastLogin
-			tt.expectedData.LastPasswordChange = user.LastPasswordChange
-			tt.expectedData.DeletedAt = user.DeletedAt
-			err := udb.Update(db, tt.expectedData)
+			user, err := udb.FindByUsername(db, tt.username)
 			assert.Equal(t, tt.expectedErr, err != nil)
 
 			if tt.expectedData != nil {
-				user = &cerebrum.User{}
-				if err := db.First(user, tt.usr.ID).Error; err != nil {
-					t.Error(err)
-				}
+				tt.expectedData.CreatedAt = user.CreatedAt
 				tt.expectedData.UpdatedAt = user.UpdatedAt
+				tt.expectedData.LastLogin = user.LastLogin
+				tt.expectedData.LastPasswordChange = user.LastPasswordChange
+				tt.expectedData.Role = superAdmin
 				assert.Equal(t, tt.expectedData, user)
+
+			}
+		})
+	}
+	db.Close()
+	if err := pool.Purge(resource); err != nil {
+		t.Fatal(fmt.Sprintf("Could not purge resource: %v", err))
+	}
+}
+
+func TestFindByToken(t *testing.T) {
+	cases := []struct {
+		name         string
+		expectedErr  bool
+		token        string
+		expectedData *cerebrum.User
+	}{
+		{
+			name:        "User does not exist",
+			expectedErr: true,
+			token:       "notExists",
+		},
+		{
+			name:  "Success",
+			token: "loginrefresh",
+			expectedData: &cerebrum.User{
+				Email:      "johndoe@mail.com",
+				FirstName:  "John",
+				LastName:   "Doe",
+				Username:   "johndoe",
+				RoleID:     cerebrum.AccessRole(100),
+				CompanyID:  1,
+				LocationID: 1,
+				Password:   "hunter2",
+				Base: cerebrum.Base{
+					Model: gorm.Model{ID: 1},
+				},
+				Token: "loginrefresh",
+			},
+		},
+	}
+
+	container := mockdb.NewMySQLDockerTestContainer(t)
+	db, pool, resource := container.DB, container.Pool, container.Resource
+
+	if err := mockdb.InsertMultiple(db, superAdmin, cases[1].expectedData); err != nil {
+		t.Error(err)
+	}
+
+	udb := store.NewUser()
+
+	for _, tt := range cases {
+		t.Run(tt.name, func(t *testing.T) {
+			user, err := udb.FindByToken(db, tt.token)
+			assert.Equal(t, tt.expectedErr, err != nil)
+
+			if tt.expectedData != nil {
+				tt.expectedData.CreatedAt = user.CreatedAt
+				tt.expectedData.UpdatedAt = user.UpdatedAt
+				tt.expectedData.LastLogin = user.LastLogin
+				tt.expectedData.LastPasswordChange = user.LastPasswordChange
+				tt.expectedData.Role = superAdmin
+				assert.Equal(t, tt.expectedData, user)
+
 			}
 		})
 	}
@@ -403,7 +449,7 @@ func TestList(t *testing.T) {
 		t.Error(err)
 	}
 
-	udb := mysqldb.NewUser()
+	udb := store.NewUser()
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
@@ -418,6 +464,83 @@ func TestList(t *testing.T) {
 					tt.expectedData[i].Role = superAdmin
 				}
 				assert.Equal(t, tt.expectedData, users)
+			}
+		})
+	}
+	db.Close()
+	if err := pool.Purge(resource); err != nil {
+		t.Fatal(fmt.Sprintf("Could not purge resource: %v", err))
+	}
+}
+
+func TestUpdate(t *testing.T) {
+	cases := []struct {
+		name         string
+		expectedErr  bool
+		usr          *cerebrum.User
+		expectedData *cerebrum.User
+	}{
+		{
+			name: "Success",
+			usr: &cerebrum.User{
+				Base: cerebrum.Base{
+					Model: gorm.Model{
+						ID: 2,
+					},
+				},
+				FirstName: "Z",
+				LastName:  "Freak",
+				Address:   "Address",
+				Phone:     "123456",
+				Mobile:    "345678",
+				Username:  "newUsername",
+			},
+			expectedData: &cerebrum.User{
+				Email:      "tomjones@mail.com",
+				FirstName:  "Z",
+				LastName:   "Freak",
+				Username:   "tomjones",
+				RoleID:     cerebrum.AccessRole(100),
+				CompanyID:  1,
+				LocationID: 1,
+				Password:   "newPass",
+				Address:    "Address",
+				Phone:      "123456",
+				Mobile:     "345678",
+				Base: cerebrum.Base{
+					Model: gorm.Model{ID: 2},
+				},
+			},
+		},
+	}
+
+	container := mockdb.NewMySQLDockerTestContainer(t)
+	db, pool, resource := container.DB, container.Pool, container.Resource
+
+	if err := mockdb.InsertMultiple(db, superAdmin, cases[0].usr); err != nil {
+		t.Error(err)
+	}
+
+	udb := store.NewUser()
+
+	for _, tt := range cases {
+		t.Run(tt.name, func(t *testing.T) {
+			user := &cerebrum.User{}
+			if err := db.First(user, tt.usr.ID).Error; err != nil {
+				t.Error(err)
+			}
+			tt.expectedData.CreatedAt = user.CreatedAt
+			tt.expectedData.LastLogin = user.LastLogin
+			tt.expectedData.LastPasswordChange = user.LastPasswordChange
+			err := udb.Update(db, tt.expectedData)
+			assert.Equal(t, tt.expectedErr, err != nil)
+			if tt.expectedData != nil {
+				user = &cerebrum.User{}
+				if err := db.First(user, tt.usr.ID).Error; err != nil {
+					t.Error(err)
+				}
+				tt.expectedData.UpdatedAt = user.UpdatedAt
+				assert.Equal(t, tt.expectedData, user)
 			}
 		})
 	}
@@ -461,7 +584,7 @@ func TestDelete(t *testing.T) {
 		t.Error(err)
 	}
 
-	udb := mysqldb.NewUser()
+	udb := store.NewUser()
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
