+++ pkg/api/user/user.go
@@ -2,10 +2,11 @@
 package user
 
 import (
+	"github.com/labstack/echo"
+
 	cerebrum "github.com/johncoleman83/cerebrum/pkg/utl/model"
 	"github.com/johncoleman83/cerebrum/pkg/utl/query"
 	"github.com/johncoleman83/cerebrum/pkg/utl/structs"
-	"github.com/labstack/echo"
 )
 
 // Create creates a new user account
